requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.1.147
multidict==6.2.0
numpy==1.26.4
openai==1.66.5
orjson==3.10.15
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.39
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
